Move one, many, many_many to new sqlite.query_helpers module
