+++ qcodes/dataset/sqlite_base.py
@@ -25,6 +25,7 @@ from qcodes.dataset.guids import generate_guid, parse_guid
 from qcodes.dataset.sqlite.connection import ConnectionPlus, atomic, \
     transaction, atomic_transaction
 from qcodes.dataset.sqlite.connection import make_connection_plus_from
+from qcodes.dataset.sqlite.query_helpers import many_many, one, many
 from qcodes.utils.types import complex_types, complex_type_union
 
 
@@ -265,56 +266,6 @@ def _adapt_complex(value: complex_type_union) -> sqlite3.Binary:
     np.save(out, np.array([value]))
     out.seek(0)
     return sqlite3.Binary(out.read())
-
-
-def one(curr: sqlite3.Cursor, column: Union[int, str]) -> Any:
-    """Get the value of one column from one row
-    Args:
-        curr: cursor to operate on
-        column: name of the column
-
-    Returns:
-        the value
-    """
-    res = curr.fetchall()
-    if len(res) > 1:
-        raise RuntimeError("Expected only one row")
-    elif len(res) == 0:
-        raise RuntimeError("Expected one row")
-    else:
-        return res[0][column]
-
-
-def many(curr: sqlite3.Cursor, *columns: str) -> List[Any]:
-    """Get the values of many columns from one row
-    Args:
-        curr: cursor to operate on
-        columns: names of the columns
-
-    Returns:
-        list of  values
-    """
-    res = curr.fetchall()
-    if len(res) > 1:
-        raise RuntimeError("Expected only one row")
-    else:
-        return [res[0][c] for c in columns]
-
-
-def many_many(curr: sqlite3.Cursor, *columns: str) -> List[List[Any]]:
-    """Get all values of many columns
-    Args:
-        curr: cursor to operate on
-        columns: names of the columns
-
-    Returns:
-        list of lists of values
-    """
-    res = curr.fetchall()
-    results = []
-    for r in res:
-        results.append([r[c] for c in columns])
-    return results
 
 
 def connect(name: str, debug: bool = False,
